+++ core/src/main/kotlin/dev/pbkit/wrp/core/startWrpServer.kt
@@ -1,9 +1,8 @@
 package dev.pbkit.wrp.core
 
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
+import kotlinx.coroutines.coroutineScope
 
-fun startWrpServer(scope: CoroutineScope, channel: WrpChannel, vararg servers: WrpServer) {
+suspend fun startWrpServer(channel: WrpChannel, vararg servers: WrpServer) {
     val availableMethods = servers
         .map { server -> server.availableMethods }
         .reduce { acc, set -> acc.union(set) }
@@ -14,7 +13,7 @@ fun startWrpServer(scope: CoroutineScope, channel: WrpChannel, vararg servers: W
         }
     }
     val host = WrpHost(channel, availableMethods)
-    scope.launch {
+    coroutineScope {
         host.listen(this).collect { request ->
             val server = methodServerMap[request.methodName] ?: return@collect
             server.handleRequest(request)
